cs7347/pyproject.toml
[tool.poetry]
name = "cs7347"
version = "0.1.0"
description = ""
authors = ["Noah Gardner <ngngardner@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.21.2"
pandas = "^1.3.3"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.4.1"
wemake-python-styleguide = "^0.15.3"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
